requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --index-url=https://pypi.tuna.tsinghua.edu.cn/simple/
#
--index-url https://pypi.tuna.tsinghua.edu.cn/simple/

asgiref==3.7.2
    # via django
async-timeout==4.0.3
    # via redis
build==1.0.3
    # via pip-tools
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via pip-tools
django==5.0
    # via
    #   django-filter
    #   django-redis
    #   django-simpleui
    #   djangorestframework
    #   djangorestframework-simplejwt
django-environ==0.11.2
    # via -r requirements.in
django-filter==23.5
    # via -r requirements.in
django-redis==5.4.0
    # via -r requirements.in
django-simpleui==2023.12.12
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
gunicorn==21.2.0
    # via -r requirements.in
idna==3.6
    # via requests
markdown==3.5.1
    # via -r requirements.in
mysqlclient==2.2.1
    # via -r requirements.in
optionaldict==0.1.2
    # via wechatpy
packaging==23.2
    # via
    #   build
    #   gunicorn
pip-tools==7.3.0
    # via -r requirements.in
pyjwt==2.8.0
    # via djangorestframework-simplejwt
pyproject-hooks==1.0.0
    # via build
python-dateutil==2.8.2
    # via wechatpy
pytz==2023.3.post1
    # via djangorestframework
redis==5.0.1
    # via django-redis
requests==2.31.0
    # via wechatpy
six==1.16.0
    # via
    #   python-dateutil
    #   wechatpy
sqlparse==0.4.4
    # via django
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
typing-extensions==4.9.0
    # via asgiref
urllib3==2.1.0
    # via requests
wechatpy==1.8.18
    # via -r requirements.in
wheel==0.42.0
    # via pip-tools
xmltodict==0.13.0
    # via wechatpy

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
